third_party/WebKit/LayoutTests/html/dialog/scrollable-after-close.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<style>
html, body { height: 100%; margin: 0; overflow: hidden; }
#scroller {
    overflow: scroll;
    position: relative;
    z-index: 0;
    width: 300px;
    height: 50%;
    left: 20px;
    top: 20px;
    border: 1px solid black;
}
</style>
<dialog></dialog>
<div id="scroller">
    <div style="height: 700px; width: 700px"></div>
</div>
<script>

description("Tests that FrameView::m_scrollableAreas is updated after " +
            "closing a modal dialog when the scroller and its ancestors " +
            "have percentage heights. (see crbug.com/633520)");

var dialog = document.querySelector("dialog");
dialog.showModal();
dialog.close();

// Force layout.
document.body.offsetWidth;

shouldBe("internals.numberOfScrollableAreas(document)", "1");

</script>
